+++ kotlin-compiler-extensions/test-gen/java/com/javiersc/kotlin/compiler/extensions/FakeDiagnosticTestGenerated.java
@@ -4,7 +4,6 @@ package com.javiersc.kotlin.compiler.extensions;
 
 import com.intellij.testFramework.TestDataPath;
 import org.jetbrains.kotlin.test.util.KtTestUtil;
-import org.jetbrains.kotlin.test.TargetBackend;
 import org.jetbrains.kotlin.test.TestMetadata;
 import org.junit.jupiter.api.Test;
 
@@ -18,6 +17,6 @@ import java.util.regex.Pattern;
 public class FakeDiagnosticTestGenerated extends AbstractFakeDiagnosticTest {
   @Test
   public void testAllFilesPresentInDiagnostics() {
-    KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/diagnostics"), Pattern.compile("^(.+)\\.kt$"), null, TargetBackend.JVM_IR, true);
+    KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/diagnostics"), Pattern.compile("^(.+)\\.kt$"), null, true);
   }
 }